Added the TopBar, but becomes even more buggy
