+++ microservices/front-end/src/components/ChangePasswordPage.tsx
@@ -3,12 +3,6 @@ import axios from "axios";
 import { useNavigate } from "react-router-dom";
 import styles from "./ChangePasswordPage.module.css";
 import { User } from "./types";
-import { AppBar } from "@mui/material";
-import Toolbar from "@mui/material/Toolbar";
-import Typography from "@mui/material/Typography";
-import Button from "@mui/material/Button";
-import Backdrop from "@mui/material/Backdrop";
-import CircularProgress from "@mui/material/CircularProgress";
 import { UserProvider, useUserContext } from "../UserContext";
 
 // Should only allow change of password if old password matches!
@@ -75,27 +69,6 @@ const ChangePasswordPage: React.FC = () => {
     }
   };
 
-  const handleQuestion = () => {
-    setNewPassword("");
-    setConfirmPassword("");
-    navigate("/question-bank");
-  };
-
-  const handleSignout = () => {
-    axios
-      .delete("/api/auth/log-out")
-      .then((response) => {
-        if (response.status === 200) {
-          setCurrentUser({} as User);
-          navigate("/login");
-        }
-      })
-      .catch((error) => {
-        console.error("Error during sign out:", error);
-        alert("Failed to sign out, please try again!");
-      });
-  };
-
   const handleCancel = () => {
     setNewPassword("");
     setConfirmPassword("");
@@ -104,21 +77,6 @@ const ChangePasswordPage: React.FC = () => {
 
   return (
     <div>
-      <div>
-        <AppBar position="static">
-          <Toolbar>
-            <Typography variant="h6" style={{ flexGrow: 1 }}>
-              HOME
-            </Typography>
-            <Button color="inherit" onClick={handleQuestion}>
-              Question
-            </Button>
-            <Button color="inherit" onClick={handleSignout}>
-              Log Out
-            </Button>
-          </Toolbar>
-        </AppBar>
-      </div>
       <div className={styles.login_container}>
         <h1>Change Password</h1>
         <div className={styles.input_field}>
